fix(SUP-46361): Transcript visibility issue with VTT captions and Label usage in Kaltura Player v7. (#234)
Issue:
When caption doesn't have label, it doesn't appear in the transcript panel.

Root cause:
Caption data is saved with key of: {language}-{label} so where it no label it will be: language-undefined.
The compare is done between this string and the {language}-{label} coming from the engine, engine added name to caption when it's empty. so there is mismatch between the values.

Fix:
remove the label from the captionKay so comparing will be only between language.